\name{cec2007_functions}
\alias{sympart}
\title{Functions from the CEC 2007 EMOA competition.}
\usage{sympart(x)
}
\description{Functions from the CEC 2007 EMOA competition.}
\value{Function value.}
\author{Olaf Mersmann \email{olafm@statistik.tu-dortmund.de}}
\keyword{optimize}
\arguments{\item{x}{Parmater vector.}
}

