\name{cec2009_functions}
\alias{UF1}
\title{Functions from the CEC 2009 EMOA competition.}
\usage{UF1(x)
UF2(x)
UF3(x)
UF4(x)
UF5(x)
UF6(x)
UF7(x)
UF8(x)
UF9(x)
UF10(x)
}
\description{Functions from the CEC 2009 EMOA competition.}
\value{\code{UF1}: Function value.

}
\author{Olaf Mersmann \email{olafm@statistik.tu-dortmund.de}}
\keyword{optimize}
\arguments{\item{x}{Parmater vector.}
}
\alias{UF2}
\alias{UF3}
\alias{UF4}
\alias{UF5}
\alias{UF6}
\alias{UF7}
\alias{UF8}
\alias{UF9}
\alias{UF10}

