\name{crowding_distance}
\alias{crowding_distance}
\title{Crowding Distance}
\usage{crowding_distance(front)
}
\description{Calculate crowding distances.}
\alias{crowding_distance}
\value{crowding distance for each function value.}
\author{Olaf Mersmann \email{olafm@statistik.tu-dortmund.de}}
\arguments{\item{front}{matrix of function values.}
}

