% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dominance_matrix}
\alias{dominance_matrix}
\title{Calculate the dominance matrix of a set of points}
\usage{
dominance_matrix(points)
}
\arguments{
\item{points}{Matrix containing points one per column.}
}
\value{
Dominance matrix
}
\description{
Calculate the dominance matrix of a set of points
}

