% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{emoa-package}
\alias{emoa-package}
\title{The EMOA package}
\description{
This package provides functions to construct evolutionary
multiobjective optimization algorithms (EMOA). The long term goal
is to also provide standard implementations of the most common
EMOA in use today.
}
\details{
Without the hard work of many researchers who have published their
source code under a liberal license, this package would not have
been possible. In alphabetical order they are

\itemize{
  \item Michael H. Buselli
  \item Wessel Dankers
  \item Carlos Fonseca
  \item Joshua Knowles
  \item Huang Ling
  \item Wudong Liu
  \item Manuel Lopez-Ibanez
  \item Luis Paquete
  \item Ponnuthurai Nagaratnam Suganthany
  \item Santosh Tiwar
  \item Qingfu Zhang
  \item Aimin Zhou
  \item Shizheng Zhaoy
}
}
\author{
Olaf Mersmann \email{olafm@statistik.tu-dortmund.de}
}
\keyword{package}

