% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{emoa_console_logger}
\alias{emoa_console_logger}
\title{console logger}
\usage{
emoa_console_logger(...)
}
\arguments{
\item{...}{passed to \code{\link{emoa_logger}}.}
}
\value{
An \code{emoa_logger} object.
}
\description{
Logger object that outputs log messages to the console
}
\details{
This is a wrapper that calls \code{emoa_logger(output=output,
...)} internally and returns that logger.
}
