% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{emoa_null_logger}
\alias{emoa_null_logger}
\title{null logger}
\usage{
emoa_null_logger(...)
}
\arguments{
\item{...}{ignored.}
}
\value{
An \code{emoa_logger} object.
}
\description{
Logger object that discards all log events.
}
