\name{pm_operator}
\alias{pm_operator}
\title{Polynomial mutation operator...}
\usage{pm_operator(n, p, lower, upper)
}
\description{Polynomial mutation operator}
\details{Returns a polynomial mutation operator with the given parameters.}
\value{Function which implements the specified mutation operator.}
\author{Olaf Mersmann \email{olafm@statistik.tu-dortmund.de}}
\arguments{\item{n}{Distance parameter mutation distribution (\eqn{\eta}{eta}).}
\item{p}{Probability of one point mutation.}
\item{lower}{Lower bounds of parameter space.}
\item{upper}{Upper bounds of parameter space.}
}

