% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicators.r
\name{unary_r2_indicator}
\alias{unary_r2_indicator}
\title{Unary R2 indicator}
\usage{
unary_r2_indicator(points, weights, ideal)
}
\arguments{
\item{points}{Matrix of points for which to calculate the indicator
value stored one per column.}

\item{weights}{Matrix of weight vectors stored one per column.}

\item{ideal}{Ideal point of true Pareto front. If omited the ideal
of \code{points} is used.}
}
\value{
Value of unary R2 indicator.
}
\description{
Unary R2 indicator
}
\author{
Olaf Mersmann \email{olafm@p-value.net}
}
