% Generated by roxygen2 (4.0.2): do not edit by hand
\name{which_points_on_edge}
\alias{which_points_on_edge}
\title{Determine which points are on the edge of a Pareto-front approximation.}
\usage{
which_points_on_edge(front)
}
\arguments{
\item{front}{Pareto-front approximation.}
}
\value{
An integer vector containing the indicies of the points
(columns) of \code{front} which are on the edge of the
Pareto-front approximation.
}
\description{
Determine which points are on the edge of a Pareto-front approximation.
}
\author{
Olaf Mersmann \email{olafm@statistik.tu-dortmund.de}
}

