% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sympart}
\alias{sympart}
\title{Functions from the CEC 2007 EMOA competition.}
\usage{
sympart(x)
}
\arguments{
\item{x}{Parmater vector.}
}
\value{
Function value.
}
\description{
Functions from the CEC 2007 EMOA competition.
}
\author{
Olaf Mersmann \email{olafm@statistik.tu-dortmund.de}
}
\keyword{optimize}

