% Generated by roxygen2 (4.0.2): do not edit by hand
\name{crowding_distance}
\alias{crowding_distance}
\title{Crowding Distance}
\usage{
crowding_distance(front)
}
\arguments{
\item{front}{matrix of function values.}
}
\value{
crowding distance for each function value.
}
\description{
Calculate crowding distances.
}
\author{
Olaf Mersmann \email{olafm@statistik.tu-dortmund.de}
}

