% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominance.r
\name{is_dominated}
\alias{is_dominated}
\alias{is_maximally_dominated}
\title{Pareto dominance checks.}
\usage{
is_dominated(points)

is_maximally_dominated(points)
}
\arguments{
\item{points}{Matrix containing points one per column.}
}
\value{
For \code{is_dominated} and \code{is_maximally_dominated}
a boolean vector and for \code{\%dominates\%} a single boolean.
}
\description{
\code{is_dominated} returns which points from a set are dominated
by another point in the set. \code{\%dominates\%} returns true if
\code{x} Pareto dominates \code{y} and
\code{is_maximally_dominated} returns TRUE for those points which
do not dominate any other points.
}
\author{
Olaf Mersmann \email{olafm@statistik.tu-dortmund.de}
}
\keyword{optimize}
