% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.r
\name{inbounds}
\alias{inbounds}
\title{Clip value to a given range}
\usage{
inbounds(x, l, u)
}
\arguments{
\item{x}{Value to clip.}

\item{l}{Lower limit.}

\item{u}{Upper limit.}
}
\value{
l if x < l, u if x > u else x.
}
\description{
Clip \eqn{x} to the interval \eqn{[l, u]}. This is useful to enforce
box constraints.
}
\author{
Olaf Mersmann \email{olafm@statistik.tu-dortmund.de}
}
