% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominance.r
\name{nondominated_points}
\alias{nondominated_points}
\title{Nondominated points}
\usage{
nondominated_points(points)
}
\arguments{
\item{points}{Matrix of points, one per column.}
}
\value{
Those points in \code{points} which are not dominated by
  another point.
}
\description{
Return those points which are not dominated by another point in
\code{points}. This is the Pareto front approximation of the
point set.
}
\author{
Olaf Mersmann \email{olafm@statistik.tu-dortmund.de}
}
\keyword{optimize}
