% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pm_control}
\alias{pm_control}
\title{Polynomial muation (PM) control parameters}
\usage{
pm_control(f, upper, lower, ..., control, default = list())
}
\arguments{
\item{f}{Multiobjectve optimization function.}

\item{upper}{Upper bounds of parameter space.}

\item{lower}{Lower bounds of parameter space.}

\item{...}{Further arguments passed to \code{f}.}

\item{control}{List of control parameters.}

\item{default}{List of default control parameters.}
}
\value{
The \code{control} list with suitably adjusted
arguments. Missing control parameters are taken from
\code{default} or, if not present there, from an internal default.
}
\description{
Control parameters:
\describe{
  \item{pm.n}{Nu parameter of PM.}
  \item{pm.p}{p parameter of PM.}
}
}
\author{
Olaf Mersmann \email{olafm@statistik.tu-dortmund.de}
}

