% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_mutation.r
\name{pm_operator}
\alias{pm_operator}
\title{Polynomial mutation operator}
\usage{
pm_operator(n, p, lower, upper)
}
\arguments{
\item{n}{Distance parameter mutation distribution (\eqn{\eta}{eta}).}

\item{p}{Probability of one point mutation.}

\item{lower}{Lower bounds of parameter space.}

\item{upper}{Upper bounds of parameter space.}
}
\value{
Function which implements the specified mutation operator.
}
\description{
Returns a polynomial mutation operator with the given parameters.
}
\author{
Olaf Mersmann \email{olafm@statistik.tu-dortmund.de}
}
