% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sb_crossover.r
\name{sbx_operator}
\alias{sbx_operator}
\title{Simulated binary crossover operator}
\usage{
sbx_operator(n, p, lower, upper)
}
\arguments{
\item{n}{Distance parameter of crossover distribution (\eqn{\eta}{eta}).}

\item{p}{Probability of one point crossover.}

\item{lower}{Lower bounds of parameter space.}

\item{upper}{Upper bounds of parameter space.}
}
\value{
Function with one parameter \code{x} which takes a matrix
  containing two sets of parameters and returns a matrix of two sets of
 parameters which resulted from the crossover operation. As with all
 \code{emoa} functions, the parameter sets are stored in the columns
 of \code{x}. \code{x} should therefore always have two columns and a
 warning will be given if it has more than two columns.
}
\description{
Returns a simulated binary crossover operator with the given parameters.
}
\seealso{
\code{\link{pm_operator}}
}
\author{
Olaf Mersmann \email{olafm@statistik.tu-dortmund.de}
}
